+++ index.html
@@ -17,13 +17,7 @@
     <script src="/socket.io/socket.io.js"></script>
     <script>
         var socket = io('localhost:8080', {
-            reconnectionDelayMax: 10000,
-            auth: {
-                token: "123"
-            },
-            query: {
-                "my-key": "my-value"
-            }
+            path: '/test'
         });
 
         socket.on('connect', (stuff) => {
